clean to kinda running
